JDK11にあわせて修正

+++ src/main/java/com/icafe4j/scripting/lexer/VariableExpression.java
@@ -39,7 +39,7 @@ class VariableExpression extends Expression {
     if (dd == null) {
       throw new ExecError(v + " is undefined");
     }
-    return (dd.doubleValue());
+    return (dd);
   }
 
   String unparse() {
